session06/labCode/requirements.txt
# Lab 6: Object Detection Requirements

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# COCO Tools
pycocotools>=2.0.7

# Visualization and Logging
matplotlib>=3.7.0
tensorboard>=2.13.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
tqdm>=4.65.0

# Optional: For better performance
# nvidia-ml-py3>=7.352.0  # GPU monitoring
# albumentations>=1.3.0   # Advanced augmentation